# rest/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Chainspace REST API endpoints",
        "title": "Chainspace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/kv/label/{label}": {
            "get": {
                "description": "get string by label",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "Retrieve a key by its label",
                "operationId": "getbyLabel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "label",
                        "name": "label",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.ObjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/api/kv/label/{label}/version-id": {
            "get": {
                "description": "get version ID by label",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "Retrieve a version ID by its label",
                "operationId": "getVersionIDbyLabel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "label",
                        "name": "label",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.VersionIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/api/kv/prefix/{prefix}": {
            "get": {
                "description": "See before",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "See before",
                "operationId": "getByPrefix",
                "parameters": [
                    {
                        "type": "string",
                        "description": "prefix",
                        "name": "prefix",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.ListObjectsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/api/kv/prefix/{prefix}/version-id": {
            "get": {
                "description": "see before",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "See before",
                "operationId": "getVersionIDbyPrefix",
                "parameters": [
                    {
                        "type": "string",
                        "description": "prefix",
                        "name": "prefix",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.ListVersionIDsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.LabelObject": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "object": {
                    "type": "object"
                },
                "versionId": {
                    "type": "string"
                }
            }
        },
        "api.LabelVersionID": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "versionId": {
                    "type": "string"
                }
            }
        },
        "api.ListObjectsResponse": {
            "type": "object",
            "properties": {
                "pairs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.LabelObject"
                    }
                }
            }
        },
        "api.ListVersionIDsResponse": {
            "type": "object",
            "properties": {
                "pairs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.LabelVersionID"
                    }
                }
            }
        },
...
        "api.ObjectResponse": {
            "type": "object",
            "properties": {
                "object": {
                    "type": "object"
                }
            }
        },
...
        "api.VersionIDResponse": {
            "type": "object",
            "properties": {
                "versionId": {
                    "type": "string"
                }
            }
        }
    }
}